requirements-mac.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
bson==0.5.10
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
pycparser==2.21
pydantic==2.6.2
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
pymongo==4.6.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.12
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
